Access Control

+++ app/Http/Controllers/PostsController.php
@@ -8,6 +8,17 @@ use App\User;
 
 class PostsController extends Controller
 {
+
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -73,6 +84,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // check for correct user
+        if(auth()->user()->id !== $post->user_id && auth()->user()->id !== 1){
+            return redirect('/posts/'.$post->id)->with('error', 'Unauthorized');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -108,6 +125,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // check for correct user
+        if(auth()->user()->id !== $post->user_id && auth()->user()->id !== 1){
+            return redirect('/posts/'.$post->id)->with('error', 'Unauthorized');
+        }
+
         $post->delete();
 
         return redirect('/posts')->with('success', 'Post Deleted');
